refactor(leetcode/.../longest_common_prefix)
Made a module
