+++ src/problems/leetcode/easy/longest_common_prefix/v1.py
@@ -1,5 +1,3 @@
-from hamcrest import assert_that, equal_to
-
 # SRC: https://leetcode.com/problems/longest-common-prefix/
 #
 # Write a function to find the longest common prefix string amongst an array of strings.
@@ -29,32 +27,3 @@ def longest_common_prefix(words: list[str]) -> str:
             return result
 
     return result
-
-
-def test(words: list[str], expected: str) -> None:
-    result = longest_common_prefix(words)
-    assert_that(result, equal_to(expected))
-    print("✅", end="")
-
-
-def run() -> None:
-    # Example1:
-    #
-    # * Input: strs = ["flower", "flow", "flight"]
-    # * Output: "fl"
-    test(["flower", "flow", "flight"], expected="fl")
-
-    # Example 2:
-    #
-    # * Input: strs = ["dog", "racecar", "car"]
-    # * Output: ""
-    # Explanation: There is no common prefix among the input strings.
-    test(["dog", "racecar", "car"], expected="")
-
-    test([""], expected="")
-
-    print("\n")
-
-
-if __name__ == "__main__":
-    run()
